Use Ubuntu, not Alpine, in hugo/build step [semver:patch] (#44)
This commit makes a (hopeful) final fix for the AsciiDoc feature
addition. The step now uses `apt-get` instead of `apk`.

The `install` step uses Alpine, but the `hugo/build` step runs on Ubuntu
18.04. So, the check for `/bin/apk` fails. Now, there is a check if the
system is a Debian-like system (making an assumption that `apt-get` will
be in the `$PATH` somewhere). If the system is a Debian derivative
operating system, like Debian, it will install `asciidoctor`.

I verified and tested this change in a local Ubuntu 18.04 container. I
verified the scriptlet works and that the packages install correctly.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

:see_no_evil: Add a colon for valid YAML

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

:superhero: Account for root privileges and privilege escalation

Per @dsayling's tip:

	https://github.com/CircleCI-Public/hugo-orb/pull/44#issuecomment-761199960

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>