Add/htmlproofer custom params to build job [semver:patch] (#46)
* :new: htmlproofer: Add new parameters for custom htmlproofer flags

This commit adds new parameters for custom flags and options to
htmlproofer. Since the htmlproofer command already has multiple options,
they only need to be exposed in the build settings.

This allows me to use htmlproofer in my workflow. Most notably, I need
the `url-ignore` option because of some internal links that will always
be broken unless they are deliberately ignored.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

* :broom: Address feedback from @dsayling in PR #45

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

* fix hugo-build command

I attempted to add a test for the job, but we'd need to include a sample
hugo project in this directory to build it and I'm no hugo expert.

Co-authored-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>