:beetle: html-proofer: Adjust error-sort flag for better error verbosity (#47)
This commit changes the `error-sort` flag to use an empty parameter in
the `html-proofer` command. This causes HTML-Proofer to show specific
error messages in CircleCI build logs, where the current behavior is to
suppress error messages.

The current default behavior makes it difficult to debug errors from
HTML-Proofer in a failing CI pipeline. Debugging requires clone the
repository and running HTML-Proofer locally to see error messages.

This is an uncommon flag to use for HTML-Proofer, and I think it does
not do exactly what it is expected to do, so the best option appears to
be setting the parameter as blank.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>